用户管理功能的修改

+++ src/main/java/com/work/work/context/UserContext.java
@@ -1,13 +1,13 @@
 package com.work.work.context;
 
 public class UserContext {
-    private static final ThreadLocal<Integer> USER_ID = new ThreadLocal<>();
+    private static final ThreadLocal<Long> USER_ID = new ThreadLocal<>();
 
-    public static void setUserId(Integer userId) {
+    public static void setUserId(Long userId) {
         USER_ID.set(userId);
     }
 
-    public static Integer getUserId() {
+    public static Long getUserId() {
         return USER_ID.get();
     }
